Refactor to allow other models to reuse DNANet_Unet functionality
